+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Support abbreviations. Closes [#498](https://github.com/fniessink/toisto/issues/498).
 - Support generating alternative answers (like accepting "it's" when the label is "is is"). Closes [#520](https://github.com/fniessink/toisto/issues/520).
 - Specifying a concept to practice with `-c/--concept` also loads related concepts.
+- Allow for specifying examples in the concept files and show those examples after quizzes.
 
 ### Removed